[master]: Working Makefile & Dockerfile.
This is the result of moving my home development environment to WSL.

As such, the Makefile and Dockerfile are updated to be the primary tool
for setting up a development environment and managing the source.

The code is modernized by use of a standardized formatting plugin.
Several linting and automatic formatting tools are added as well.

The documentation generation and reformatting is now completely
scripted. This results in a slightly different format for the final
docs.